Port tests to new API
